backend/backend-python/django_backend/user_services/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.3.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
